Uso di transient per evitare di inserire campi non necessari nel json e calcolo del danno delle armi per caratteristica

+++ src/main/java/it/unicam/cs/mpgc/rpg130397/elements/abstractelements/Characteristics.java
@@ -27,7 +27,7 @@ public class Characteristics {
         INTELLIGENCE
     }
 
-    public float getCharacteristic(CharacteristicType type) {
+    public int getCharacteristic(CharacteristicType type) {
         return characteristics.get(type);
     }
